# apps/studio/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Studio API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "summary": "Get all settings overrides for hydration",
        "tags": [
          "settings"
        ],
        "responses": {
          "200": {
            "description": "List of settings-overrides documents"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Upsert settings for a scope",
        "tags": [
          "settings"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scope": {
                    "type": "string",
                    "enum": [
                      "app",
                      "workspace",
                      "editor"
                    ]
                  },
                  "scopeId": {
                    "type": "string",
                    "nullable": true
                  },
                  "settings": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created or updated settings document"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/structured-output": {
      "post": {
        "summary": "Get structured JSON output from prompt using a named or custom schema",
        "tags": [
          "ai"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string"
                  },
                  "schemaName": {
                    "type": "string"
                  },
                  "schema": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parsed JSON per schema"
          },
          "400": {
            "description": "prompt required or invalid schema"
          },
          "503": {
            "description": "OpenRouter not configured"
          }
        }
      }
    },
    "/api/video-docs/{id}": {
      "get": {
        "summary": "Get video doc by ID",
        "tags": [
          "video-docs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video doc document"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Update video doc",
        "tags": [
          "video-docs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "doc": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated video doc document"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/video-docs": {
      "get": {
        "summary": "List video docs",
        "tags": [
          "video-docs"
        ],
        "responses": {
          "200": {
            "description": "List of video doc documents"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create video doc",
        "tags": [
          "video-docs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "doc": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created video doc document"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
...
}